use incoming pull request events only as a hint of what is going on
Like the push events, pull request events are sometimes lost. This
causes pull requests to be kept, that are actually closed (and
mayhem whenever we need to restart a daemon)

This commit implements the same scheme than for the push events:
incoming pull request events are used as a hint that "something"
happened on the pull request. The actual processing is done on
the state of the pull request as fetched from the repository APIs.