refactor the github watcher to emit the full state of the objects that might have changed
In the current architecture, the events were passed to the buildconf
manager for processing. This led to a lot of little bugs with major
impact (usually, build loops).

This refactors the interaction between the watcher and the buildconf
manager to only pass the objects that might have changed (inferred
from the events), freshly taken from GitHub itself. Working with
the full state is a lot easier on both sides.