feat: make the strategy to choose the test branch configurable
I personally would like to force using the merge commit. Before I bit
the bullet and looked for it, the bug related to mergeability on github
made me waste a LOT of time checking why pull requests won't build,
to finally realize it's because things that were on master were actually
not included.

At least, if we always use 'merge' (except for drafts), I'm going to
get a clear update error that the merge commit does not exist.