fix: detection of merge commit on GitHub
In some conditions, GitHub will not update the merge commit unless
explicitly queried, either by opening the pull request page on the
Web UI, or doing a GET on the specific pull request endpoint. In
the latter case, the mergeable attribute indicates that the merge
commit is being updated, which allows us to wait for it before we
return and let the daemon continue.