fix: do not query the Git services when resolving dependencies
We resolve all pull requests in a single spot, so we can actually
resolve all dependencies there once and for all.

Until now it was not such a big deal since we were using faraday
cache, but now that we disabled it to workaround the merge issues
on github, it has a big impact. Essentially, we are doing one query
per pull request in #pull_requests, plus one per pull request in the
PR dependency tree, including the root. Duplicated if a given PR is
more than once in a dependency tree.

That's a lot of requests.